requirements.txt
aniso8601==9.0.1
async-timeout==4.0.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
Deprecated==1.2.13
Flask==2.1.2
Flask-RESTful==0.3.9
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.22.4
packaging==21.3
pandas==1.4.2
pymongo==4.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
redis==4.3.3
requests==2.27.1
six==1.16.0
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.14.1
yara-python==4.2.0
zipp==3.8.0
